deps: Add Ninja to the toolchain
So we can stop bloating our repo with prebuilt binaries.

+++ releng/setup-env.sh
@@ -245,7 +245,9 @@ if ! grep -Eq "^$toolchain_version\$" "$FRIDA_TOOLROOT/VERSION.txt" 2>/dev/null;
   ) > "$vala_wrapper"
   chmod 755 "$vala_wrapper"
 
-  ln -s "$FRIDA_ROOT/releng/ninja-$build_os_arch" "$FRIDA_TOOLROOT/bin/ninja"
+  if [ ! -e "$FRIDA_TOOLROOT/bin/ninja" ]; then
+    ln -s "$FRIDA_ROOT/releng/ninja-$build_os_arch" "$FRIDA_TOOLROOT/bin/ninja"
+  fi
   ln -s "$FRIDA_ROOT/releng/frida-resource-compiler-$build_os_arch" "$FRIDA_TOOLROOT/bin/frida-resource-compiler"
 else
   detect_vala_api_version
